requirements.txt
streamlit 
neo4j 
chromadb 
langchain 
langchain-community 
langchain-text-splitters 
groq 
python-dotenv 
sentence-transformers